--- 2015-F/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  2015-F
-//
-//  Created by 木下　和巳 on 11/17/15.
-//  Copyright © 2015 ie-ModelingAndDesign. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
